requirements.txt
# Core dependencies
PyYAML>=6.0
python-dateutil>=2.8.0

# LLM integration (optional)
openai>=1.0.0

# Optional dependencies for enhanced functionality
# pandas>=1.3.0  # For DataFrame support
# numpy>=1.21.0  # For numerical computations
# matplotlib>=3.5.0  # For visualization
# seaborn>=0.11.0  # For statistical visualization

# Development dependencies (uncomment for development)
# pytest>=7.0.0
# pytest-cov>=4.0.0
# black>=22.0.0
# flake8>=5.0.0